### 2022-05-17 #4401 @editorialbot
[PRE REVIEW]: DSO++: Modern reimplementation of Direct Sparse Odometry
**Submitting author:** @Solonets (<a href="http://orcid.org/0000-0002-3469-7489">Sergei Solonets</a>)
**Repository:** https://github.com/RoadlyInc/DSOPP
**Branch with paper.md** (empty if default branch): feature/joss_paper
**Version:** v 1.0
