### 2022-05-17 #4401 @Kevin-Mattheus-Moerman
@Solonets I'm the AEiC on call this week for JOSS. I am here to help get over initial hurdles and to assign a handling editor when ready. Can you check those potentially missing DOI's :point_up: , note you can call `@editorialbot check references` to repeat the check.
