+++ joss/paper.bib
@@ -6,6 +6,7 @@
   number={3},
   pages={611--625},
   year={2017},
+  doi={10.1109/TPAMI.2017.2658577},
   publisher={IEEE}
 }
 
@@ -14,6 +15,7 @@
   author={Wang, Rui and Schworer, Martin and Cremers, Daniel},
   booktitle={Proceedings of the IEEE International Conference on Computer Vision},
   pages={3903--3911},
+  doi={10.48550/arXiv.1708.07878}
   year={2017}
 }
 
@@ -22,6 +24,7 @@
   author={Schubert, David and Demmel, Nikolaus and Usenko, Vladyslav and Stuckler, Jorg and Cremers, Daniel},
   booktitle={Proceedings of the European Conference on Computer Vision (ECCV)},
   pages={682--697},
+  doi={10.1007/978-3-030-01237-3_42}
   year={2018}
 }
 
@@ -31,6 +34,7 @@
   booktitle={2018 IEEE/RSJ International Conference on Intelligent Robots and Systems (IROS)},
   pages={2198--2204},
   year={2018},
+  doi={10.1109/IROS.2018.8593376}
   organization={IEEE}
 }
 
@@ -45,10 +49,11 @@
   doi={10.1109/LRA.2021.3140129}}
 
 
-  @inproceedings{yang2020d3vo,
+@inproceedings{yang2020d3vo,
   title={D3vo: Deep depth, deep pose and deep uncertainty for monocular visual odometry},
   author={Yang, Nan and Stumberg, Lukas von and Wang, Rui and Cremers, Daniel},
   booktitle={Proceedings of the IEEE/CVF Conference on Computer Vision and Pattern Recognition},
   pages={1281--1292},
+  doi={10.48550/arXiv.2003.01060}
   year={2020}
 }
