+++ joss/paper.bib
@@ -15,7 +15,7 @@
   author={Wang, Rui and Schworer, Martin and Cremers, Daniel},
   booktitle={Proceedings of the IEEE International Conference on Computer Vision},
   pages={3903--3911},
-  doi={10.48550/arXiv.1708.07878}
+  doi={10.48550/arXiv.1708.07878},
   year={2017}
 }
 
@@ -24,7 +24,7 @@
   author={Schubert, David and Demmel, Nikolaus and Usenko, Vladyslav and Stuckler, Jorg and Cremers, Daniel},
   booktitle={Proceedings of the European Conference on Computer Vision (ECCV)},
   pages={682--697},
-  doi={10.1007/978-3-030-01237-3_42}
+  doi={10.1007/978-3-030-01237-3_42},
   year={2018}
 }
 
@@ -34,7 +34,7 @@
   booktitle={2018 IEEE/RSJ International Conference on Intelligent Robots and Systems (IROS)},
   pages={2198--2204},
   year={2018},
-  doi={10.1109/IROS.2018.8593376}
+  doi={10.1109/IROS.2018.8593376},
   organization={IEEE}
 }
 
@@ -46,7 +46,7 @@
   volume={7},
   number={2},
   pages={1408-1415},
-  doi={10.1109/LRA.2021.3140129}}
+  doi={10.1109/LRA.2021.3140129}},
 
 
 @inproceedings{yang2020d3vo,
@@ -54,6 +54,6 @@
   author={Yang, Nan and Stumberg, Lukas von and Wang, Rui and Cremers, Daniel},
   booktitle={Proceedings of the IEEE/CVF Conference on Computer Vision and Pattern Recognition},
   pages={1281--1292},
-  doi={10.48550/arXiv.2003.01060}
+  doi={10.48550/arXiv.2003.01060},
   year={2020}
 }
