### 2022-05-18 #4401 @Solonets
Hi, @Kevin-Mattheus-Moerman , thank you much for your help. I fixed the paper, the way it should satisfy the formal requirements. It's still tiny, Even though I could be wordier if needed, I prefer to stay as concise as possible. We provide only a reimplementation, so there is not much to report.
