+++ joss/paper.md
@@ -34,10 +34,13 @@ bibliography: paper.bib
 
 # Summary
 
-We present DSO++, a modern reimplementation of one of the most well-known simultaneous localization and mapping (SLAM) engines. As well as its ancestor our system produces a trajectory and a sparse point cloud from a video sequence. It acts on original image and directly estimate motion by minimizing the photometric error. Compared to the original code, our version is covered with tests and modular. We also designed the code with the possibility of being extended especially concentrating on the examples of the follow-up papers. 
+We present DSO++, a modern reimplementation of one of the most well-known Simultaneous Localization And Mapping (SLAM) engines: Direct Sparse Odometry (DSO) [@engel2017direct]. Similar to its ancestor, our system produces a trajectory and a sparse point cloud from a video sequence. It acts on original image data and directly estimates motion by minimizing the photometric error. Compared to the original code, our version is covered with tests and is modular. We also designed the code with the possibility of being easily extended especially concentrating on the proposed extensions (various camera model formulations, multi-sensor setup, etc.). Another contribution is extensive testing of the functionality.
 
 # Statement of need
 
-Direct simultaneous localization and mapping (SLAM) is an active research topic in the computer vision community. One of the first working examples of direct SLAM is Direct Sparse Odometry(DSO) by Engel et al [@engel2017direct]. Since the release of the DSO code, researchers have proposed several extensions to the original method [@schubert2018direct] [@wang2017stereo] [@gao2018ldso] [@9669044] [@yang2020d3vo]. The original DSO code was not designed with a presumption of continuous development and, therefore, does not follow many good software principles. Consequently, follow-up works inherit the mistakes in the code as well as poor software practices. To facilitate the research in the field we propose our reimplementation of the DSO designed to be easily modifiable and extendable. Another contribution is extensive testing of the functionality and supporting mathematical hypotheses.
+Direct simultaneous localization and mapping is an active research topic in the computer vision community. One of the first working examples of direct SLAM is DSO. Since the release of the DSO code, researchers have proposed several extensions to the original method, introducing such features as rolling shutter motion modeling [@schubert2018direct], stereo [@wang2017stereo], deep learning aided depth inference [@yang2020d3vo] [@gnnet] and other [@9669044] [@gao2018ldso] [@omni]. Almost all of the extensions are released by the same computer vision lab due to the complexity of the original DSO code. The code was not designed with a presumption of continuous development and, therefore, is hard to maintain and extend, especially by outside researchers from other groups. 
+
+The research purpose of this software is twofold. Firstly, it provides researchers with implemented and tested modules for direct SLAM: photometric bundle adjustment, feature metric bundle adjustment, marginalization, etc., which could be useful for any subsequent work as well as student projects. Secondly, similar to any other structure from motion algorithm it can be directly used by researchers in the area of photogrammetry, 3D reconstruction, differentiable rendering, and other similar areas to estimate camera motion and approximate scene structure for further enhancement. 
+
 
 # References
