+++ joss/paper.bib
@@ -57,3 +57,60 @@
   doi={10.48550/arXiv.2003.01060},
   year={2020}
 }
+
+@article{gnnet,
+  author    = {Lukas von Stumberg and
+               Patrick Wenzel and
+               Qadeer Khan and
+               Daniel Cremers},
+  title     = {GN-Net: The Gauss-Newton Loss for Deep Direct {SLAM}},
+  journal   = {CoRR},
+  volume    = {abs/1904.11932},
+  year      = {2019},
+  url       = {http://arxiv.org/abs/1904.11932},
+  eprinttype = {arXiv},
+  eprint    = {1904.11932},
+  timestamp = {Thu, 02 May 2019 15:13:44 +0200},
+  biburl    = {https://dblp.org/rec/journals/corr/abs-1904-11932.bib},
+  doi = {10.48550/arXiv.1904.11932}, 
+  bibsource = {dblp computer science bibliography, https://dblp.org}
+}
+
+@article{other1,
+  author    = {Hidenobu Matsuki and
+               Lukas von Stumberg and
+               Vladyslav Usenko and
+               J{\"{o}}rg St{\"{u}}ckler and
+               Daniel Cremers},
+  title     = {Omnidirectional {DSO:} Direct Sparse Odometry with Fisheye Cameras},
+  journal   = {CoRR},
+  volume    = {abs/1808.02775},
+  year      = {2018},
+  url       = {http://arxiv.org/abs/1808.02775},
+  eprinttype = {arXiv},
+  eprint    = {1808.02775},
+  doi = {10.1109/LRA.2018.2855443},
+  timestamp = {Mon, 01 Jun 2020 15:30:41 +0200},
+  biburl    = {https://dblp.org/rec/journals/corr/abs-1808-02775.bib},
+  bibsource = {dblp computer science bibliography, https://dblp.org}
+}
+
+
+@article{omni,
+  author    = {Hidenobu Matsuki and
+               Lukas von Stumberg and
+               Vladyslav Usenko and
+               J{\"{o}}rg St{\"{u}}ckler and
+               Daniel Cremers},
+  title     = {Omnidirectional {DSO:} Direct Sparse Odometry with Fisheye Cameras},
+  journal   = {CoRR},
+  volume    = {abs/1808.02775},
+  year      = {2018},
+  url       = {http://arxiv.org/abs/1808.02775},
+  eprinttype = {arXiv},
+  eprint    = {1808.02775},
+  doi = {10.1109/LRA.2018.2855443},
+  timestamp = {Mon, 01 Jun 2020 15:30:41 +0200},
+  biburl    = {https://dblp.org/rec/journals/corr/abs-1808-02775.bib},
+  bibsource = {dblp computer science bibliography, https://dblp.org}
+}
