+++ joss/paper.md
@@ -38,9 +38,9 @@ We present DSO++, a modern reimplementation of one of the most well-known Simult
 
 # Statement of need
 
-Direct simultaneous localization and mapping is an active research topic in the computer vision community. One of the first working examples of direct SLAM is DSO. Since the release of the DSO code, researchers have proposed several extensions to the original method, introducing such features as rolling shutter motion modeling [@schubert2018direct], stereo [@wang2017stereo], deep learning aided depth inference [@yang2020d3vo] [@gnnet] and other [@9669044] [@gao2018ldso] [@omni]. Almost all of the extensions are released by the same computer vision lab due to the complexity of the original DSO code. The code was not designed with a presumption of continuous development and, therefore, is hard to maintain and extend, especially by outside researchers from other groups. 
+Direct simultaneous localization and mapping is an active research topic in the computer vision community. One of the first working examples of direct SLAM is DSO. Since the release of the DSO code, researchers have proposed several extensions to the original method, introducing such features as rolling shutter motion modeling [@schubert2018direct], stereo [@wang2017stereo], deep learning aided depth inference [@yang2020d3vo] [@gnnet] and other [@9669044] [@gao2018ldso] [@omni]. Almost all of the extensions are released by the same computer vision lab, most likely due to the complexity of the original DSO code. The code was not designed with a presumption of continuous development and, therefore, is hard to maintain and extend, especially by outside researchers from other groups. 
 
-The research purpose of this software is twofold. Firstly, it provides researchers with implemented and tested modules for direct SLAM: photometric bundle adjustment, feature metric bundle adjustment, marginalization, etc., which could be useful for any subsequent work as well as student projects. Secondly, similar to any other Structure From Motion algorithm it can be directly used by researchers in the area of photogrammetry, 3D reconstruction, differentiable rendering, and other similar areas to estimate camera motion and approximate scene structure for further enhancement. 
+The research purpose of this software is twofold. Firstly, it provides researchers with implemented and tested modules for direct SLAM: photometric bundle adjustment, feature metric bundle adjustment, marginalization, etc., which could be useful for any subsequent work as well as student projects. Secondly, similar to any other Structure From Motion algorithms it can be directly used by researchers in the area of photogrammetry, 3D reconstruction, differentiable rendering, and other similar areas. In those areas, researchers need estimated camera motion and/or approximate scene structure for further enhancement. Today it is common to use camera poses obtained from the indirect SLAM algorithms, which are shown to be less accurate compared to direct ones [@engel2017direct].  
 
 
 # References
