### 2022-05-19 #4401 @Solonets
@Kevin-Mattheus-Moerman thank you for such detailed feedback, I fixed the grammar mistakes you found in the first part.

> Consider adding more detail in the paper on some of the claims/statements, e.g.
>
> * [ ]  About `...examples of the follow-up papers...`, it is unclear what these follow-up papers or the examples are, please briefly clarify.

I deleted this.

> * [ ]  `Consequently, follow-up works inherit the mistakes in the code as well as poor software practices.` What "mistakes" are these, and also what "poor software practices"? Can you mention/cite examples? Can you cite follow-up works?

I tried to hedge from the discussion about the quality of the original code, it's indeed a hard claim to support.  I do not want to put any evidence of this into the text, it would look a bit harsh.

> * [ ]  What was the original algorithm implemented in? Also C++ or a different language?

it's also in C++, do you think I should mention this in a text?
> * [ ]  You mention `Another contribution is extensive testing of the functionality and supporting mathematical hypotheses`.Can you elaborate briefly on the novel testing and what it covers? Can you also comment on what mathematical hypotheses are now supported or how?

I removed this.  I was just talking about some correctness proofs for analytical gradients, marginalization, etc. I guess It's ussualy part of the regular testing, and I shouldn't state it separately.
> * [ ]  In the statement of need you express that your re-implementation is more extensible and perhaps is implemented using best practices in software development (a cleaner, clearer, better implementation is sounds like). However it is not clear what the `research purpose of the software` is. This is another point the statement of need section should make. Can you help sketch a quick picture of how this work will be or is used in research (you can cite some studies or describe how it has been or will be used) ?

I added some discussion on it. Please take a look.
