### 2022-06-02 75f23df
fix marginalization
