+++ src/energy/n_point_solvers/include/energy/n_point_solvers/pure_rotation_estimator.hpp
@@ -1,6 +1,8 @@
 #ifndef DSOPP_TRACKER_MONOCULAR_INITIALIZER_PURE_ROTATION_ESTIMATOR_HPP_
 #define DSOPP_TRACKER_MONOCULAR_INITIALIZER_PURE_ROTATION_ESTIMATOR_HPP_
 
+#include <optional>
+
 #include <Eigen/Dense>
 #include <sophus/se3.hpp>
 
