+++ src/common/timestamp_storage/include/common/timestamp_storage/timestamp_storage.hpp
@@ -3,6 +3,7 @@
 
 #include <cstdint>
 #include <map>
+#include <optional>
 
 #include "common/time/time.hpp"
 
