+++ cmake/modules/opencv.cmake
@@ -1,57 +1,83 @@
 set(CONTRIB_URL https://github.com/opencv/opencv_contrib.git)
 set(CONTRIB_VERSION 3b5a55876fe0502418a9d9fb7d388c40f2a626b1)
 
-set(contrib_INCLUDE_DIR ${CMAKE_CURRENT_BINARY_DIR}/3rd_party/src/opencv_contrib_external/)
-set(contrib_LIBS ${CMAKE_CURRENT_BINARY_DIR}/3rd_party/src/opencv_contrib_external/modules)
+set(contrib_INCLUDE_DIR
+    ${CMAKE_CURRENT_BINARY_DIR}/3rd_party/src/opencv_contrib_external/)
+set(contrib_LIBS
+    ${CMAKE_CURRENT_BINARY_DIR}/3rd_party/src/opencv_contrib_external/modules)
 
-if (NOT EXISTS ${contrib_LIBS})
-    externalproject_add(opencv_contrib_external
-            GIT_REPOSITORY ${CONTRIB_URL}
-            GIT_TAG        ${CONTRIB_VERSION}
-            CONFIGURE_COMMAND ""
-            CMAKE_COMMAND ""
-            BUILD_COMMAND ""
-            INSTALL_COMMAND ""
-            TEST_COMMAND ""
-            PREFIX 3rd_party
-            EXCLUDE_FROM_ALL 1)
-    file(MAKE_DIRECTORY ${contrib_INCLUDE_DIR})
+if(NOT EXISTS ${contrib_LIBS})
+  ExternalProject_Add(
+    opencv_contrib_external
+    GIT_REPOSITORY ${CONTRIB_URL}
+    GIT_TAG ${CONTRIB_VERSION}
+    CONFIGURE_COMMAND ""
+    CMAKE_COMMAND ""
+    BUILD_COMMAND ""
+    INSTALL_COMMAND ""
+    TEST_COMMAND ""
+    PREFIX 3rd_party
+    EXCLUDE_FROM_ALL 1)
+  file(MAKE_DIRECTORY ${contrib_INCLUDE_DIR})
 endif()
 
 add_library(opencv_contrib INTERFACE IMPORTED GLOBAL)
 add_dependencies(opencv_contrib opencv_contrib_external)
 
-
 set(URL https://github.com/opencv/opencv.git)
 set(VERSION 9aa647068b2eba4a34462927b1878353dfd3df69)
 
-set(opencv_INCLUDE_DIR ${CMAKE_CURRENT_BINARY_DIR}/3rd_party/src/opencv_external-build/install/include/opencv4)
-set(opencv_LIBS ${CMAKE_CURRENT_BINARY_DIR}/3rd_party/src/opencv_external-build/install/lib)
+set(opencv_INCLUDE_DIR
+    ${CMAKE_CURRENT_BINARY_DIR}/3rd_party/src/opencv_external-build/install/include/opencv4
+)
+set(opencv_LIBS
+    ${CMAKE_CURRENT_BINARY_DIR}/3rd_party/src/opencv_external-build/install/lib)
 
 get_target_property(glog_INCLUDE_DIRS glog INTERFACE_INCLUDE_DIRECTORIES)
 get_property(glog_lib GLOBAL PROPERTY glog_path)
 
-if (NOT EXISTS ${opencv_LIBS})
-    externalproject_add(opencv_external
-            DEPENDS opencv_contrib glog
-            GIT_REPOSITORY ${URL}
-            GIT_TAG ${VERSION}
-            CMAKE_ARGS
-            -DCMAKE_BUILD_TYPE=Release
-            -DINSTALL_C_EXAMPLES=OFF
[…]
