+++ cmake/modules/glog.cmake
@@ -9,9 +9,12 @@ if (NOT EXISTS ${glog_LIBS})
             GIT_TAG        ${VERSION}
             CMAKE_ARGS
                 -DCMAKE_BUILD_TYPE=Release
-                -DCMAKE_CXX_FLAGS=-march=native -std=c++20
+                -DCMAKE_CXX_STANDARD=20
+                -DCMAKE_CXX_STANDARD_REQUIRED=ON
+                -DCMAKE_CXX_FLAGS=-march=native
                 -DCMAKE_INSTALL_PREFIX=./install
                 -DBUILD_TESTING=OFF
+                -DWITH_UNWIND=OFF
             TEST_COMMAND ""
             PREFIX 3rd_party
             EXCLUDE_FROM_ALL 1

[…]
