### 2022-06-14 bcf6254
Fix build on modern compilers (#4)
