+++ src/feature_based_slam/tracker/internal/feature_based_slam/tracker/feature_frame_from_landmarks.hpp
@@ -1,6 +1,7 @@
 #ifndef DSOPP_FEATURE_BASED_SLAM_TRACKER_CREATE_LANDMARKS_HPP
 #define DSOPP_FEATURE_BASED_SLAM_TRACKER_CREATE_LANDMARKS_HPP
 
+#include <stddef.h>
 #include <deque>
 #include <vector>
 
