+++ src/feature_based_slam/tracker/internal/feature_based_slam/tracker/add_new_projections_to_landmarks.hpp
@@ -1,6 +1,7 @@
 #ifndef DSOPP_FEATURE_BASED_SLAM_TRACKER_ADD_NEW_PROJECTIONS_TO_LANDMARKS_HPP
 #define DSOPP_FEATURE_BASED_SLAM_TRACKER_ADD_NEW_PROJECTIONS_TO_LANDMARKS_HPP
 
+#include <stddef.h>
 #include <vector>
 
 namespace dsopp {
