+++ src/tracker/keyframe_strategy/src/mean_square_optical_flow_and_rmse_keyframe_strategy.cpp
@@ -40,7 +40,7 @@ bool MeanSquareOpticalFlowAndRmseKeyframeStrategy<Motion>::needNewKeyframe(
 }
 
 template <energy::motion::Motion Motion>
-MeanSquareOpticalFlowAndRmseKeyframeStrategy<Motion>::~MeanSquareOpticalFlowAndRmseKeyframeStrategy<Motion>() = default;
+MeanSquareOpticalFlowAndRmseKeyframeStrategy<Motion>::~MeanSquareOpticalFlowAndRmseKeyframeStrategy() = default;
 
 template class MeanSquareOpticalFlowAndRmseKeyframeStrategy<energy::motion::SE3<Precision>>;
 
