+++ src/sensors/sensors_builder/include/sensors/sensors_builder/camera_fabric.hpp
@@ -5,6 +5,7 @@
 #include <any>
 #include <map>
 #include <memory>
+#include <optional>
 #include <string>
 #include <vector>
 
