### 2022-06-17 #4401 @hugoledoux
you cannot do it yourself, but you can recommend me names and I'll approach them!

We have a list but I tried most having computer vision as keywords... It can be academics, researchers in governmental organisation or companies. But (obviously) not your direct colleagues or people you worked in the past with...
