+++ test/test/sensors/camera/test_camera.cpp
@@ -20,15 +20,17 @@ namespace dsopp {
 namespace sensors {
 
 TEST(testCamera, oneCamera) {
-  cv::Mat image(100, 100, CV_8UC3, cv::Scalar(255, 255, 255));
-  image.at<cv::Vec3b>(50, 50) = cv::Vec3b(0, 0, 0);
+  auto vignetting = calibration::vignetting::create(TEST_DATA_DIR "track30seconds/vignetting.png");
+  auto cols = std::max(vignetting.cols, 100);
+  auto rows = std::max(vignetting.rows, 100);
+  cv::Mat image(rows, cols, CV_8UC3, cv::Scalar(255, 255, 255));
+  image.at<cv::Vec3b>(rows / 2, cols / 2) = cv::Vec3b(0, 0, 0);
   auto provider = std::make_unique<providers::MockCameraProvider>();
   auto *data_frame = new providers::CameraDataFrame(0, std::move(image), time(std::chrono::milliseconds(0)));
   EXPECT_CALL(*provider, nextFrameProxy()).WillOnce(::testing::Return(data_frame));
   auto calib = calibration::CameraCalibration(Eigen::Vector2<Precision>::Zero(), Eigen::Vector4<Precision>::Zero(),
                                               energy::model::ModelType::kPinholeCamera);
   auto photo_calib = calibration::photometric_calibration::create(TEST_DATA_DIR "track30seconds/pcalib.txt");
-  auto vignetting = calibration::vignetting::create(TEST_DATA_DIR "track30seconds/vignetting.png");
 
   Eigen::Vector2<int> im_size = calib.image_size().cast<int>();
   auto camera_mask = calibration::CameraMask(im_size.y(), im_size.x());
