+++ pydsopp/track/track.py
@@ -49,7 +49,6 @@ class Track:
     """
     Class for exctract data from track storage
     """
-
     def __init__(self, track_storage):
         """
         Arguments:
