### 2022-06-20 80e76c4
fix vignetting size
