+++ src/sensors/sensors_builder/src/camera_fabric.cpp
@@ -163,10 +163,6 @@ std::optional<calibration::CameraSettings> createCameraSettings(
 
   camera_mask = calibration::CameraMask(runMaskTransformers(transformers, undistorter.undistort(camera_mask->data())));
 
-  if (!vignetting.empty()) {
-    vignetting = cv::Mat(runImageTransformers(transformers, vignetting));
-  }
-
   // read sematincs legend
   std::unique_ptr<semantics::SemanticLegend> semantic_legend = nullptr;
 
