+++ src/features/include/features/camera/pixel_data_frame_extractor.hpp
@@ -20,12 +20,9 @@ class PixelDataFrameExtractor {
  public:
   /**
    * creates extractor
-   * @param photometric_calibration gamma correction to apply to image
-   * @param undistorted_vignetting vignetting to apply to image
    * @param pyramid_levels number of pyramids to produce
    */
-  PixelDataFrameExtractor(const std::array<Precision, 256>& photometric_calibration,
-                          const cv::Mat& undistorted_vignetting, const size_t& pyramid_levels);
+  PixelDataFrameExtractor(const size_t& pyramid_levels);
   /**
    * extracts pyramids from image
    * @param frame iimage to extract image
@@ -34,10 +31,6 @@ class PixelDataFrameExtractor {
   std::unique_ptr<PixelDataFrame> extract(const cv::Mat& frame) const;
 
  private:
-  /** photometric_calibration gamma correction to apply to image */
-  const std::array<Precision, 256>& photometric_calibration_;
-  /** vignetting to apply to image */
-  const cv::Mat& undistorted_vignetting_;
   /** number of pyramids to produce */
   const size_t pyramid_levels_;
 };
