### 2022-06-21 e5e8315
Revert "make sobel more stable (#6)"
This reverts commit d980ed9d5e2b948a73812b916ee32ca3d6770fa0.
