+++ src/features/src/sobel_tracking_features_extractor.cpp
@@ -1,6 +1,7 @@
 #include "features/camera/sobel_tracking_features_extractor.hpp"
 
 #include <optional>
+#include <random>
 
 #include "features/camera/pixel_data_frame.hpp"
 #include "features/camera/pixel_map.hpp"
@@ -22,35 +23,26 @@ namespace {
  * @param current_threshold current threshold which needs to be modified
  * @return new threshold
  */
-Precision calculateThreshold(int num_pixels, int desired_points, int found_points, Precision current_threshold) {
-  return current_threshold *
-         static_cast<Precision>(std::log(num_pixels / desired_points) / std::log(num_pixels / found_points));
+int calculateThreshold(int num_pixels, int desired_points, int found_points, int current_threshold) {
+  return static_cast<int>(current_threshold * std::log(num_pixels / desired_points) /
+                          std::log(num_pixels / found_points));
 }
 
 /**
  * calculate q-quantile of gradient norms
  *
  * @param grad_norm matrix of gradient norm squared for all pixels
- * @param mask mask for valid points
  * @param q quantile level
  * @return q-quantile of gradient norms
  */
-int16_t quantile(cv::Mat &grad_norm, const sensors::calibration::CameraMask &mask, Precision q) {
-  std::vector<int16_t> grad_norms;
-  grad_norms.reserve(static_cast<size_t>(grad_norm.rows * grad_norm.cols));
-  for (int y = 0; y < grad_norm.rows; ++y) {
-    for (int x = 0; x < grad_norm.cols; ++x) {
-      if (mask.valid(x, y)) {
-        grad_norms.push_back(grad_norm.at<int16_t>(y, x));
-      }
-    }
-  }
+int16_t quantile(cv::Mat &grad_norm, Precision q) {
+  std::vector<int16_t> grad_norms(grad_norm.begin<int16_t>(), grad_norm.end<int16_t>());
   std::nth_element(grad_norms.begin(),
                    grad_norms.begin() + static_cast<long>(static_cast<Precision>(grad_norms.size()) * q),
                    grad_norms.end());
 
   return static_cast<int16_t>(grad_norms[static_cast<size_t>(static_cast<Precision>(grad_norms.size()) * q)]);
-}  // namespace
+}
 
 /**
  * find point with high gradient in window_size X window_size block
@@ -61,58 +53,26 @@ int16_t quantile(cv::Mat &grad_norm, const sensors::calibration::CameraMask &mas
  * @param window_start_x x coord of start point in common matrix
  * @param window_start_y y coord of start point in common matrix
  * @param grad_norm_threshold threshold for gradient norm
- * @param weight_of_mean weight of mean in threshold
  * @return first point with enough high gradient if it exist
  */
 std::optional<Eigen::Matrix<Precision, 2, 1>> findPointInWindow(cv::Mat &grad_norm,
                                                                 const sensors::calibration::CameraMask &mask,
                                                                 int window_size, int window_start_x, int window_start_y,
-                                                                Precision grad_norm_threshold,
-                                                                Precision weight_of_mean) {
-  Precision average =
-      static_cast<Precision>(cv::mean(grad_norm(cv::Range(window_start_y, window_start_y + window_size),
-                                                cv::Range(window_start_x, window_start_x + window_size)))[0]);
-  Precision threshold = std::sqrt((weight_of_mean * average * average + grad_norm_threshold * grad_norm_threshold) /
-                                  (1 + weight_of_mean));
+                                                                int grad_norm_threshold) {
[…]
