### 2022-06-21 c52bf41
Revert "make sobel more stable" (#7)
Revert "make sobel more stable (#6)"

This reverts commit d980ed9d5e2b948a73812b916ee32ca3d6770fa0.
