+++ src/features/src/pixel_data_frame_extractor.cpp
@@ -4,9 +4,14 @@
 
 namespace dsopp {
 namespace features {
-PixelDataFrameExtractor::PixelDataFrameExtractor(const size_t& pyramid_levels) : pyramid_levels_(pyramid_levels) {}
+PixelDataFrameExtractor::PixelDataFrameExtractor(const std::array<Precision, 256>& photometric_calibration,
+                                                 const cv::Mat& undistorted_vignetting, const size_t& pyramid_levels)
+    : photometric_calibration_(photometric_calibration),
+      undistorted_vignetting_(undistorted_vignetting),
+      pyramid_levels_(pyramid_levels) {}
 std::unique_ptr<PixelDataFrame> PixelDataFrameExtractor::extract(const cv::Mat& frame) const {
-  return std::make_unique<features::PixelDataFrame>(frame, pyramid_levels_);
+  return std::make_unique<features::PixelDataFrame>(frame, photometric_calibration_, undistorted_vignetting_,
+                                                    pyramid_levels_);
 }
 }  // namespace features
 }  // namespace dsopp
