+++ src/features/src/camera_features.cpp
@@ -13,8 +13,7 @@
 
 namespace dsopp {
 namespace features {
-CameraFeatures::CameraFeatures(const size_t frame_id, cv::Mat&& raw_image, cv::Mat&& photocorrected_undistorted_image,
-                               const time time,
+CameraFeatures::CameraFeatures(const size_t frame_id, cv::Mat&& raw_image, const time time,
                                const std::vector<sensors::calibration::CameraMask>& pyramid_of_static_masks,
                                features::TrackingFeaturesExtractor& tracking_features_extractor,
                                const features::PixelDataFrameExtractor& pixel_data_frame_extractor,
@@ -22,20 +21,19 @@ CameraFeatures::CameraFeatures(const size_t frame_id, cv::Mat&& raw_image, cv::M
                                const semantics::SemanticFilter* semantic_filter)
     : frame_id_(frame_id),
       raw_image_(std::move(raw_image)),
-      frame_data_(std::move(photocorrected_undistorted_image)),
       time_(time),
       pyramid_of_static_masks_(pyramid_of_static_masks),
       tracking_features_extractor_(tracking_features_extractor),
       pixel_data_frame_extractor_(pixel_data_frame_extractor),
       semantics_data_(std::move(semantics_data)),
       semantic_filter_(semantic_filter) {
-  cv::cvtColor(raw_image_, raw_frame_data_, cv::COLOR_BGR2GRAY);
+  cv::cvtColor(raw_image_, frame_data_, cv::COLOR_BGR2GRAY);
 }
 
 size_t CameraFeatures::id() const { return frame_id_; }
 const features::TrackingFeaturesFrame& CameraFeatures::tracking() {
   if (tracking_features_ == nullptr) {
-    tracking_features_ = tracking_features_extractor_.extract(raw_frame_data_, pyramidOfMasks()[0]);
+    tracking_features_ = tracking_features_extractor_.extract(frame_data_, pyramidOfMasks()[0]);
   }
   return *tracking_features_;
 }
@@ -48,7 +46,7 @@ const features::PixelDataFrame& CameraFeatures::pixelData() {
 
 const cv::Mat& CameraFeatures::image() const { return raw_image_; }
 
-const cv::Mat& CameraFeatures::frameData() const { return raw_frame_data_; }
+const cv::Mat& CameraFeatures::frameData() const { return frame_data_; }
 
 std::unique_ptr<cv::Mat> CameraFeatures::moveSemanticsData() { return std::move(semantics_data_); }
 
