### 2022-06-29 #4521 @Mirmix
Thanks for the efforts. Here is my review of the paper.

    What we mean by research software
    JOSS publishes articles about research software. This definition includes software that: solves complex modeling  problems in a scientific context (physics, mathematics, biology, medicine, social science, neuroscience,
    engineering); supports the functioning of research instruments or the execution of research experiments;
    extracts knowledge from large data sets; offers a mathematical library, or similar.
    While useful for many areas of research, pre-trained machine learning models
    and notebooks are not in-scope for JOSS.
    ***
    Substantial scholarly effort
    JOSS publishes articles about software that represent substantial scholarly effort on the part of the authors. Your software should be a significant contribution to the available open source software that either enables
    some new research challenges to be addressed or makes addressing research challenges significantly better (e.g., faster, easier, simpler).

The main contribution of the authors is to make DSO more modular and have extensive testing of the functionality of already existing software. So here, we can't talk much about improvement in time/accuracy performance, thus it may seem like more of an engineering effort than a scholarly effort. Which is totally fine in JOSS.

However, I believe the user should see clear reasons which can be in the documentation/user-friendliness of the software that can attract the user to use this software rather than using an original and well-known implementation of DSO.

    Some factors that may be considered by editors and reviewers when judging effort include:
        * Age of software (is this a well-established software project) / length of commit history.
        * Number of commits.
        * Number of authors.
        * Total lines of code (LOC). Submissions under 1000 LOC will usually be flagged, those under 300 LOC will be desk rejected.
        * Whether the software has already been cited in academic papers.
        * Whether the software is sufficiently useful that it is likely to be cited by your peer group.

At the current stage, I think the repository is badly documented. It lacks documenting lots of part like installation (each step guidance), dataset formats, how dataset and code organized, what some files(.sh) are for and etc. I have faced multiple issues while trying to install the repository. Personally, I think the overall repository structure/tree is too modular and contains too many files/folders that makes it harder to track than the original DSO implementation.

Original DSO has much better documentation with an end-to-end installation manual, which makes it easy-to-use and easy-to-install (and/though it does not use docker). The structure of the code is more user-friendly and easy to grasp.

The repository has some ".sh" files without any documentation and explanation for them. I would like to also see end-to-end installation documentation.

    The paper should be between 250-1000 words. Authors submitting papers significantly longer than 1000 words may be asked to reduce the length of their paper.

Excluding the (intertext) references the paper is roughly around 350 words. It fits with the guideline, but I find it a bit short. I would be glad if the author wrote more about the "delta" of their work, against the original implementation. I would like to also see time performance and maybe memory consumption which shows that the current reimplementation does not hit the performance.

### Datasets issues
Your code use some test data from TUM MONO, consider citing it in README.

Your code also uses some test data by downloading it using "download_test_data.sh".
You should also put information about how the calibration files and other files are organized, what are the fields in each *.csv, *.txt, *.gt files are.

# Some installation issues
## Docker
This software uses docker with Ubuntu of 18.04. If we are talking about the modern implementation of DSO, I would personally expect the usage of a more recent Ubuntu version and more recent packages. There is no also proper manual, thus the user is expected to have knowledge about how to use docker, and how to install this software which is not usually the cases among the target audience (students).

You may also consider to put docker image in docker hub server, for simple docker pull option.

I have faced many issues while installation (using naive docker installation):
At the first step I naively run docker build and docker run at the location of Dockerfile

```
@DSOPP/docker/ubuntu$ sudo docker build -t dsopp -f Dockerfile .
sudo docker run -it dsopp
```

There was docker build error due to lcov_1.15 was not found, so consider solving it with lcov_1.16 version in the Dockerfile(line 79):

    RUN wget http://ftp.de.debian.org/debian/pool/main/l/lcov/lcov_1.16.orig.tar.xz \
        && tar xvf lcov_1.16.orig.tar.xz \
        && cd lcov-1.16 \
        && make install \
        && cd .. \
        && rm -rf lcov*

You need to also consider adding the following packages to install(download_test_data.sh uses them, so needs them)

    pip install gdown
    apt-get install zip

## General repository build
When I try to git clone/cmake/make (build repository in standard way)  the repository as shown in the README it gives another error:

    mkdir build;
    cd build;
    cmake .. -DCMAKE_CXX_COMPILER=g++-10 -DCMAKE_BUILD_TYPE=Release;
    make -j;

```
-- Configuring done
-- Generating done
-- Build files have been written to: /app/build/DSOPP/build
root@17731041aa19:/app/build/DSOPP/build# make
Scanning dependencies of target check_format
--- /app/build/DSOPP/pydsopp/track/frame.py	(original)
+++ /app/build/DSOPP/pydsopp/track/frame.py	(reformatted)
@@ -23,12 +23,10 @@
     """
     Frame in the world
     """
-
     class AttachedFrame:
         """
         Frame attached to a keyframe (Frame)
         """
-
         def __init__(self, id_, timestamp, odometry_t_keyframe_frame):
             self.id = id_
             self.timestamp = timestamp
--- /app/build/DSOPP/pydsopp/track/camera_calibration.py	(original)
+++ /app/build/DSOPP/pydsopp/track/camera_calibration.py	(reformatted)
@@ -11,7 +11,6 @@
     """
     Pinhole camera model
     """
-
     def __init__(self, focal_x, focal_y, center_x, center_y):
         """
         Intialize Pinhole camera via 3x3 calibration matrix
@@ -62,7 +61,6 @@
     """
     Camera calibration
     """
-
     def __init__(self, intrinsics, image_size, model_type):
         if model_type.name == 'pinhole':
             self.model = Pinhole(*intrinsics[:4])
--- /app/build/DSOPP/pydsopp/track/track.py	(original)
+++ /app/build/DSOPP/pydsopp/track/track.py	(reformatted)
@@ -49,7 +49,6 @@
     """
     Class for exctract data from track storage
     """
-
     def __init__(self, track_storage):
         """
         Arguments:
--- /app/build/DSOPP/pydsopp/track/export/colmap_track_export.py	(original)
+++ /app/build/DSOPP/pydsopp/track/export/colmap_track_export.py	(reformatted)
@@ -12,7 +12,6 @@
     """
     Internal class for building colmap data, representation of 3D point with index and 2D projection indices
     """
-
     @dataclass
     class Projection:
         frame_index: int
@@ -32,7 +31,6 @@
     """
     Internal class for building colmap data
     """
-
     @dataclass
     class Point2D:
         pixel: np.array
--- /app/build/DSOPP/pydsopp/transformations/motion.py	(original)
+++ /app/build/DSOPP/pydsopp/transformations/motion.py	(reformatted)
@@ -6,7 +6,6 @@

 class Motion:
     """ 3 dimensional group of similarity transformations """
-
     def __init__(self, q, t):
         """
         Construct `Motion` from q,t or sympy matrix
CMakeFiles/check_format.dir/build.make:74: recipe for target 'CMakeFiles/check_format' failed
make[2]: *** [CMakeFiles/check_format] Error 1
CMakeFiles/Makefile2:2326: recipe for target 'CMakeFiles/check_format.dir/all' failed
make[1]: *** [CMakeFiles/check_format.dir/all] Error 2
Makefile:111: recipe for target 'all' failed
make: *** [all] Error 2

```
## Python library only build

In README there is another error/typo

    Original: git clone --depth 1 https://github.com/RoadAR/DSOPP
    Should be: git clone --depth 1 https://github.com/RoadlyInc/DSOPP

I have managed to run "python library only build"

    git clone --depth 1 https://github.com/RoadlyInc/DSOPP
    mkdir build && cd build
    cmake ../ -DBUILD_DOC=ON -DCHECK_FORMAT=ON
    make pydsopp_lib

After running python scripts I was getting the following error message:

    FileNotFoundError: [Errno 2] No such file or directory: '*.bin'

Maybe this is due to previous issues but ctest was also failing:

    The following tests FAILED:
        1 - test_track2tum_exporter (Not Run)
        2 - test_config_loader (Not Run)
        3 - test_pinhole_camera_model (Not Run)
        4 - test_atan_camera_model (Not Run)
        5 - test_ios_camera_model (Not Run)
        6 - test_simple_radial (Not Run)
        7 - test_division_camera_model (Not Run)
        8 - test_tum_fov_model (Not Run)
        9 - test_reprojects (Not Run)
        10 - test_projector (Not Run)
        11 - test_se3_motion (Not Run)
        12 - test_generalized_epipolar_line (Not Run)
        13 - test_triangulation (Not Run)
        14 - test_essential_matrix (Not Run)
        15 - test_affine_brightness (Not Run)
        16 - test_ceres_pose_alignment (Not Run)
        17 - test_photometric_bundle_adjustment (Not Run)
        18 - test_local_parameterization_s2 (Not Run)
        19 - test_ceres_pose_alignment_noised (Not Run)
        20 - test_analytical_diff (Not Run)
        21 - test_linear_system (Not Run)
        22 - test_incremental_solver (Not Run)
        23 - test_ceres_camera_optimization (Not Run)
        24 - test_geometric_bundle_adjustment (Not Run)
        25 - test_tracking_features_extractor (Not Run)
        26 - test_visualizer (Not Run)
        27 - test_image_folder_provider (Not Run)
        28 - test_camera (Not Run)
        29 - test_undistorter (Not Run)
        30 - test_get_camera_model (Not Run)
        31 - test_camera_mask (Not Run)
        32 - test_image_video_provider (Not Run)
        33 - test_npy_folder_provider (Not Run)
        34 - test_camera_transforms (Not Run)
        35 - test_camera_settings (Not Run)
        36 - test_frequency_keyframe_strategy (Not Run)
        37 - test_depth_estimation (Not Run)
        38 - test_landmarks_activator (Not Run)
        39 - test_reference_frame_depth_map (Not Run)
        40 - test_track_storage (Not Run)
        41 - test_track_frames (Not Run)
        42 - test_marginalization (Not Run)
        43 - test_fbs_feature_matcher (Not Run)
        44 - test_fbs_tracker (Not Run)
        45 - test_pydsopp_json_exporter (Failed)
        46 - test_extract_tracking_features (Failed)
        47 - test_python_reader (Failed)
    Errors while running CTest

## Conclusions
Document the package and installation extensively.

Your title claim

    DSO++: Modern reimplementation of Direct Sparse Odometry

So I would expect it to use more recent OS (ubuntu) and package (dependency) versions

You claim that:

    It provides researchers with implemented and tested modules for direct SLAM: photometric bundle adjustment, feature metric bundle
    adjustment, marginalization, etc., which could be useful for any subsequent work as well as student projects.

If it is a goal, I would expect the repository to provide **tutorial codes**, and **end-to-end easy-to-use/install manual** for an average student interested in Multi-View Geometry. If anyhow you see "pydsopp" as possible tutorial examples, then you need to well comment and document them.

You also claim that:

    Secondly, similar to any other Structure From Motion algorithms it can be directly used by researchers in the area of photogrammetry, 3D reconstruction, differentiable rendering, and other similar areas.

If it is a goal, it should bring some kind of simplicity over already existing photogrammetry software/libraries like colmap/openMVG/openGV and should be also well documented.

I believe if you are **reimplementing/refactoring** the already existing library, at least it should be at the same level of well-documented, user and developer friendly as already existing software. Furthermore explaining "delta" and necessity for this new library well is very important. Unfortunately with the current stage of your repository and work, I believe the original DSO is much more appealing and easy to use.
