### 2022-07-05 e813a75
Fix build on g++ 12
