+++ cmake/modules/gflags.cmake
@@ -26,5 +26,6 @@ endif()
 add_library(gflags INTERFACE IMPORTED GLOBAL)
 add_dependencies(gflags gflags_external)
 file(MAKE_DIRECTORY ${gflags_INCLUDE_DIR})
-target_include_directories(gflags INTERFACE ${glog_INCLUDE_DIR})
+target_include_directories(gflags INTERFACE ${gflags_INCLUDE_DIR})
+set_property(GLOBAL PROPERTY gflags_path ${gflags_LIBS})
 target_link_libraries(gflags INTERFACE ${gflags_LIBS})
