### 2022-07-05 be712cc
Delete cout test
