+++ docker/run_local_CI.sh
@@ -67,7 +67,7 @@ docker system prune -f
 
 # BUILDING DOCKER
 
-docker image build -t dsopp:test $script_directory/$docker_OS
+docker image build -t dsopp:test --build-arg USER_ID=$(id -u) --build-arg GROUP_ID=$(id -g) $script_directory/$docker_OS
 
 # RUNNING CI PIPELINE
 
