### 2022-07-06 cde7d5f
Fix build on modern compilers (g++12) (#9)
