+++ src/output/visualizer/internal/visualizer/buffer_storage.hpp
@@ -3,6 +3,7 @@
 
 #include <map>
 
+#include <GL/glew.h>
 #include <pangolin/pangolin.h>
 
 namespace dsopp {
