### 2022-07-06 44f1fec
Basic Installation and usage docs (#10)
