+++ README.md
@@ -2,30 +2,173 @@
 
 Reimplementation of **Direct Sparse Odometry**, *J. Engel, V. Koltun, D. Cremers*, In arXiv:1607.02565, 2016 (https://github.com/JakobEngel/dso) 
 
-if used in a publication please consider citing: **Upcomming publication**
+if used in a publication please consider citing: **Upcoming publication**
 
 for commercial use and additional functionality contact Roadly (en@road.ly)
 
-## Installation
+## Building DSOPP
 
-### dependencies
+First of all, you need to clone this repository
 
-to install dependecies follow a corresponding docker file:
-  [Ubuntu](docker/ubuntu/Dockerfile)
+```
+git clone git@github.com:RoadlyInc/DSOPP.git
+cd DSOPP
+```
 
+### Dependencies installation
 
-### Build
+<details>
+<summary>Without Docker [Not recommended]</summary>
+
+Be sure that `python3` is installed on your machine.
+
+#### Using python virtual evironment [recommended]
+
+It is recommended to exploit python virtual environments for each external project not to mess up your system's python dependecies.
+To install `virtualenv` package:
+
+```
+pip3 install virtualenv 
+```
+
+To create a virtual environment run the following command
+
+```
+python3 -m venv dsopp_venv
+```
+
+And to activate:
+
+```
+source dsopp_venv/bin/activate
+```
+
+Now depending on your shell settings, you might see `(e) dsopp_venv` in your terminal line.
+
+#### Installing python dependencies
+
+From the root of this repository run:
+
+```
+pip3 -r requirements.txt
+```
+
+#### system dependencies
+
+If you are on ubuntu or debian system:
+
+```
+apt install cmake libboost-all-dev graphviz doxygen libgl1-mesa-dev libglew-dev 
+apt install git-lfs libprotobuf-dev protobuf-compiler libhdf5-dev libfreetype6-dev
+apt install libtbb-dev openexr libavcodec-dev libavformat-dev libswscale-dev 
+apt install libavutil-dev clang sed gawk clang-10 zlib1g clang-format-10 ninja-build
+apt install wget ffmpeg libeigen3-dev libsuitesparse-dev lcov python2
+```
+
+Note: This package names are valid for ubuntu 18.04 and 20.04.
+
+For installation on other systems, please find the alternative name of each package above in your package manager.
+
+</details>
+
+<details>
+<summary>Docker [recommended]</summary>
+
+Docker is a great way to run a code within specific environment on any machine.
+To install docker on ubuntu or debian:
+
+```
+sudo apt install docker
+```
+
+After installation [docker group should be created and added to user](https://docs.docker.com/engine/install/linux-postinstall/).
+
+```
+sudo groupadd docker
+sudo usermod -aG docker $USER
+```
+
+Do not forget to log out after running the above commands.
+Now `cd` to `docker/ubuntu/` and run the following command to build an image
+
+```
+docker image build -t dsopp:main --build-arg USER_ID=(id−u) −−build−arg GROUPID=(id -g) .
+```
+
+To connect to `dsopp:main` image and use gui docker should be added to xhost
+
+```
+xhost local:docker
+```
+
+And following command to open shell inside `dsopp:main` image in the current directory:
+
+```
+docker run --rm -it -v /tmp/.X11-unix:/tmp/.X11-unix -e HOME=HOME−eDISPLAY=DISPLAY -w (pwd)−vHOME:$HOME --device=/dev/dri:/dev/dri -it dsopp:main bash
+```
+
+</details>
+
+#### Build
+
[…]
