+++ docker/ubuntu/Dockerfile
@@ -1,4 +1,4 @@
-FROM ubuntu:18.04
+FROM ubuntu:20.04
 
 ENV DEBIAN_FRONTEND=noninteractive
 
@@ -88,7 +88,14 @@ RUN wget http://ftp.de.debian.org/debian/pool/main/l/lcov/lcov_1.16.orig.tar.xz 
 RUN cd /usr/include && ln -sf eigen3/Eigen Eigen &&  ln -sf eigen3/unsupported unsupported
 
 
-RUN mkdir /app
-WORKDIR /app
+WORKDIR /
+
+ARG USER_ID
+ARG GROUP_ID
+
+RUN addgroup --gid $GROUP_ID user
+RUN adduser --disabled-password --gecos '' --uid $USER_ID --gid $GROUP_ID user
+
+USER user
 
 CMD ["/bin/bash"]
