### 2022-07-07 #4521 @Solonets
Dear @Mirmix, thank you for your detailed review. We added much more extensive documentation and fixed some of the problems you mention. Would you be ready to check it out once again?
