### 2022-07-09 96af2a7
readme: missing install in pip3 install -r requierements.txt
