### 2022-07-09 #4521 @jhidalgocarrio
Thanks to the Authors for this contribution. I reviewed the installation process

### Building DSOPP

#### Installing python dependencies

`pip3 -r requirements.txt`  this command is missing the install option. PR [here](https://github.com/RoadlyInc/DSOPP/pull/11)

#### Build

Using the `virtualenv` on Ubuntu 18.04
```
mkdir build;
cd build;
cmake .. -DCMAKE_BUILD_TYPE=Release;
make -j;
```
build stops here:
```
[ 19%] Completed 'eigen_external'
[ 19%] Built target eigen_external
[ 98%] Linking CXX static library ../../../lib/libgmock_main.a
[ 98%] Built target gmock_main
[100%] Linking CXX executable yaml-cpp-tests
[100%] Built target yaml-cpp-tests
[ 20%] No install step for 'yaml-cpp_external'
[ 20%] No test step for 'yaml-cpp_external'
[ 20%] Completed 'yaml-cpp_external'
[ 20%] Built target yaml-cpp_external
Makefile:100: recipe for target 'all' failed
make: *** [all] Error 2
```

Scrolling up I see two errors:

```
[  7%] Running python protocol buffer compiler on semantic_legend.proto
c++: error: unrecognized command line option ‘-fconcepts-diagnostics-depth=2’; did you mean ‘-fno-diagnostics-color’?
src/common/time/CMakeFiles/time.dir/build.make:75: recipe for target 'src/common/time/CMakeFiles/time.dir/src/time.cpp.o' failed
make[2]: *** [src/common/time/CMakeFiles/time.dir/src/time.cpp.o] Error 1
CMakeFiles/Makefile2:2204: recipe for target 'src/common/time/CMakeFiles/time.dir/all' failed
make[1]: *** [src/common/time/CMakeFiles/time.dir/all] Error 2
```

```
-- Found Threads: TRUE
[ 16%] Built target gflags_external
-- Configuring done
CMake Error in src/CMakeLists.txt:
  Target "benchmark" requires the language dialect "CXX20" (with compiler
  extensions), but CMake does not know the compile flags to use to enable it.

CMake Error in src/CMakeLists.txt:
  Target "benchmark_main" requires the language dialect "CXX20" (with
  compiler extensions), but CMake does not know the compile flags to use to
  enable it.
```
