+++ README.md
@@ -132,7 +132,7 @@ Do not forget to log out after running the above commands.
 Now `cd` to `docker/ubuntu/` and run the following command to build an image
 
 ```
-docker image build -t dsopp:main --build-arg USER_ID=(id−u) −−build−arg GROUPID=(id -g) .
+docker image build -t dsopp:main --build-arg USER_ID=$(id−u) −−build−arg GROUPID=$(id -g) .
 ```
 
 To connect to `dsopp:main` image and use gui docker should be added to xhost
@@ -144,7 +144,7 @@ xhost local:docker
 And following command to open shell inside `dsopp:main` image in the current directory:
 
 ```
-docker run --rm -it -v /tmp/.X11-unix:/tmp/.X11-unix -e HOME=HOME−eDISPLAY=DISPLAY -w (pwd) −v HOME:$HOME --device=/dev/dri:/dev/dri -it dsopp:main bash
+docker run --rm -it -v /tmp/.X11-unix:/tmp/.X11-unix -e HOME=$HOME −e DISPLAY=$DISPLAY -w $(pwd) −v HOME:$HOME --device=/dev/dri:/dev/dri -it dsopp:main bash
 ```
 
 </details>
