### 2022-07-12 09bfa02
Add g++-10 and cmake sections to Readme.md (#12)
