### 2022-07-12 #4521 @DaniilSinitsyn
@jhidalgocarrio
Thank you for your effort to build our code!
And thank you for spotting our mistake in pip command.
Unfortunately, I forgot to add a section about using recent versions of g++ compiler and cmake...

I added those sections there. And comments about the possibility of facing a building issues through non-docker way.
