### 2022-07-12 4bdcc7e
readme: missing install in pip3 install -r requierements.txt (#11)
Co-authored-by: jhidalgocarrio <havyhidago@gmail.com>
