### 2022-07-13 8c57681
Add nerf util and pydsopp readme
