### 2022-07-15 #4521 @hugoledoux
@Solonets @DaniilSinitsyn the comment above from @Mirmix is very relevant (for the paper):

    I would be glad if the author wrote more about the "delta" of their work, against the original implementation. "

Since your work is a rewrite and improvements over the original DSO, the users of the original code should be informed about what has changed and why you think it's better.
