### 2022-07-20 #4521 @Mirmix
Thanks for the updates and for addressing some of my aforementioned issues. Although there are still a few unaddressed issues are still there, to not repeat myself here (and since the main updates are in installation) I will be mainly focused on Docker-based installation issues that are still there.

I tried to install the repo using docker. As I mentioned before some packages were forgotten (*zip*, *unzip* and *gdown*) so to solve it I created a [pull request](https://github.com/RoadlyInc/DSOPP/pull/14),  please consider merging it.

I have run ctest and Tests **#45** and **#46** are failed

```bash
      Start  1: test_track2tum_exporter
 1/47 Test  #1: test_track2tum_exporter ..............   Passed    0.05 sec
      Start  2: test_config_loader
 2/47 Test  #2: test_config_loader ...................   Passed    5.88 sec
      Start  3: test_pinhole_camera_model
 3/47 Test  #3: test_pinhole_camera_model ............   Passed    6.72 sec
      Start  4: test_atan_camera_model
 4/47 Test  #4: test_atan_camera_model ...............   Passed    0.02 sec
      Start  5: test_ios_camera_model
 5/47 Test  #5: test_ios_camera_model ................   Passed    0.00 sec
      Start  6: test_simple_radial
 6/47 Test  #6: test_simple_radial ...................   Passed    0.00 sec
      Start  7: test_division_camera_model
 7/47 Test  #7: test_division_camera_model ...........   Passed    0.00 sec
      Start  8: test_tum_fov_model
 8/47 Test  #8: test_tum_fov_model ...................   Passed    0.00 sec
      Start  9: test_reprojects
 9/47 Test  #9: test_reprojects ......................   Passed    0.00 sec
      Start 10: test_projector
10/47 Test #10: test_projector .......................   Passed    0.00 sec
      Start 11: test_se3_motion
11/47 Test #11: test_se3_motion ......................   Passed    0.00 sec
      Start 12: test_generalized_epipolar_line
12/47 Test #12: test_generalized_epipolar_line .......   Passed    0.04 sec
      Start 13: test_triangulation
13/47 Test #13: test_triangulation ...................   Passed    0.00 sec
      Start 14: test_essential_matrix
14/47 Test #14: test_essential_matrix ................   Passed    0.00 sec
      Start 15: test_affine_brightness
15/47 Test #15: test_affine_brightness ...............   Passed    0.59 sec
      Start 16: test_ceres_pose_alignment
16/47 Test #16: test_ceres_pose_alignment ............   Passed    3.29 sec
      Start 17: test_photometric_bundle_adjustment
17/47 Test #17: test_photometric_bundle_adjustment ...   Passed    4.93 sec
      Start 18: test_local_parameterization_s2
18/47 Test #18: test_local_parameterization_s2 .......   Passed    0.00 sec
      Start 19: test_ceres_pose_alignment_noised
19/47 Test #19: test_ceres_pose_alignment_noised .....   Passed    1.21 sec
      Start 20: test_analytical_diff
20/47 Test #20: test_analytical_diff .................   Passed    0.74 sec
      Start 21: test_linear_system
21/47 Test #21: test_linear_system ...................   Passed    8.34 sec
      Start 22: test_incremental_solver
22/47 Test #22: test_incremental_solver ..............   Passed    1.74 sec
      Start 23: test_ceres_camera_optimization
23/47 Test #23: test_ceres_camera_optimization .......   Passed    8.32 sec
      Start 24: test_geometric_bundle_adjustment
24/47 Test #24: test_geometric_bundle_adjustment .....   Passed    0.01 sec
      Start 25: test_tracking_features_extractor
25/47 Test #25: test_tracking_features_extractor .....   Passed    0.67 sec
      Start 26: test_visualizer
26/47 Test #26: test_visualizer ......................   Passed    0.01 sec
      Start 27: test_image_folder_provider
27/47 Test #27: test_image_folder_provider ...........   Passed    0.02 sec
      Start 28: test_camera
28/47 Test #28: test_camera ..........................   Passed    0.02 sec
      Start 29: test_undistorter
29/47 Test #29: test_undistorter .....................   Passed    0.76 sec
      Start 30: test_get_camera_model
30/47 Test #30: test_get_camera_model ................   Passed    0.00 sec
      Start 31: test_camera_mask
31/47 Test #31: test_camera_mask .....................   Passed    0.98 sec
      Start 32: test_image_video_provider
32/47 Test #32: test_image_video_provider ............   Passed    5.52 sec
      Start 33: test_npy_folder_provider
33/47 Test #33: test_npy_folder_provider .............   Passed    3.33 sec
      Start 34: test_camera_transforms
34/47 Test #34: test_camera_transforms ...............   Passed    0.06 sec
      Start 35: test_camera_settings
35/47 Test #35: test_camera_settings .................   Passed    0.03 sec
      Start 36: test_frequency_keyframe_strategy
36/47 Test #36: test_frequency_keyframe_strategy .....   Passed    0.00 sec
      Start 37: test_depth_estimation
37/47 Test #37: test_depth_estimation ................   Passed    7.14 sec
      Start 38: test_landmarks_activator
38/47 Test #38: test_landmarks_activator .............   Passed    9.07 sec
      Start 39: test_reference_frame_depth_map
39/47 Test #39: test_reference_frame_depth_map .......   Passed    1.48 sec
      Start 40: test_track_storage
40/47 Test #40: test_track_storage ...................   Passed    0.23 sec
      Start 41: test_track_frames
41/47 Test #41: test_track_frames ....................   Passed    0.03 sec
      Start 42: test_marginalization
42/47 Test #42: test_marginalization .................   Passed    0.00 sec
      Start 43: test_fbs_feature_matcher
43/47 Test #43: test_fbs_feature_matcher .............   Passed    0.76 sec
      Start 44: test_fbs_tracker
44/47 Test #44: test_fbs_tracker .....................   Passed   33.03 sec
      Start 45: test_pydsopp_json_exporter
45/47 Test #45: test_pydsopp_json_exporter ...........***Failed    0.09 sec
      Start 46: test_extract_tracking_features
46/47 Test #46: test_extract_tracking_features .......***Failed    0.02 sec
      Start 47: test_python_reader
47/47 Test #47: test_python_reader ...................   Passed    2.00 sec

96% tests passed, 2 tests failed out of 47

Label Time Summary:
Coverage    =  10.35 sec*proc (25 tests)

Total Test time (real) = 107.21 sec

The following tests FAILED:
	 45 - test_pydsopp_json_exporter (Failed)
	 46 - test_extract_tracking_features (Failed)
Errors while running CTest
```

Another thing that would make me glad if it will be resolved, is I saw you have tried to remove the password prompts inside the docker with lines below (from Dockerfile), which is good.

```bash
ARG USER_ID
ARG GROUP_ID

RUN addgroup --gid $GROUP_ID user
RUN adduser --disabled-password --gecos '' --uid $USER_ID --gid $GROUP_ID user
```

But unfortunately, it was not working as expected, when i was trying to run
```bash
(sudo) apt-get install XXX
```
it was asking me for a password. I have tried my Ubuntu password and empty string it did not work.

```bash
(base) user@169e5297fc95:~/Workspace/DSOPP/build$ sudo apt-get install smth
[sudo] password for user:
Sorry, try again.
[sudo] password for user:
Sorry, try again.
[sudo] password for user:
sudo: 3 incorrect password attempts
(base) user@169e5297fc95:
```

Maybe it can be somehow helpful, when I am printing `echo $USER` I am getting an empty string, and `whoami` returns "user"
```
(base) user@169e5297fc95:~/Workspace/DSOPP/build$ echo $USER

(base) user@169e5297fc95:~/Workspace/DSOPP/build$ whoami
user
```
