+++ joss/paper.md
@@ -42,5 +42,18 @@ Direct simultaneous localization and mapping is an active research topic in the 
 
 The research purpose of this software is twofold. Firstly, it provides researchers with implemented and tested modules for direct SLAM: photometric bundle adjustment, feature metric bundle adjustment, marginalization, etc., which could be useful for any subsequent work as well as student projects. Secondly, similar to any other Structure From Motion algorithms it can be directly used by researchers in the area of photogrammetry, 3D reconstruction, differentiable rendering, and other similar areas. In those areas, researchers need estimated camera motion and/or approximate scene structure for further enhancement. Today it is common to use camera poses obtained from the indirect SLAM algorithms, which are shown to be less accurate compared to direct ones [@engel2017direct].  
 
+# Contribution
+
+Compared to the original implementation we propose following improvements:
+
+- *Modularity*. Original DSO consists of a few source files with extensive code duplication. For example, the code responsible for ray reprojection can be found in 30-50 different places. It not only makes the development process error prone but also makes the code extremely hard to modify. However, for some followup papers, this chunk of code was the only change (rolling shutter DSO[@schubert2018direct], omniderectional DSO[@omni]). In our implementation, code is organized in a way that those kinds of contributions would require only a single modification. 
+
+- *Tests*. Every module in our implementation is covered with a test. Tests not only insure the validity of the code but also declare the essence of this module and show their usage. There are not tests in the original DSO.
+
+- *Extensibility*. functionality of every followup research have been or can be developed within our framework. It can be turned on or off dynamically based on the configuration file. It prevents duplication of code among several repositories which insures every contribution affects previous and following contributions. 
+
+
+
+
 
 # References
