### 2022-07-26 258288f
Add sudo to docker, fix (?) python test (#15)
