### 2022-08-03 #4521 @Solonets
Hi @jhidalgocarrio, thank you for your effort. There was an undefined behavior in CMake which caused no visualization during the run. I fixed it.

Regarding DCMAKE_CXX_STANDARD maybe we could continue this discussion in PR if you shall open one.
