### 2022-08-04 bf68ff9
fix memory leak (#16)
