### 2022-08-08 #4521 @Mirmix
Hi, thanks for the updates.
I ran the code with docker, Pangolin GUI was not working with the default docker run command (most probably due to the fact that the default network driver for docker is [bridge](https://docs.docker.com/network/)), after explicitly setting the network driver to "host" I was able to run the DSOPP on test sequences with GUI on my _Ubuntu20.04_. Thus please consider this [pull request](https://github.com/RoadlyInc/DSOPP/pull/18).

Like other visual SLAM/Odometry algorithms DSO aims to run in real-time (by the robotic community usually for visual monoslam to be considered real-time minimum frame rate supposed to be at least 25 or 30 Hz).  In the GUI, I was observing a much slower frame rate. Can you please share the main reason for that? It will be even better if you share with us the time performance comparison **with the same default settings and hardware on the same dataset (maybe on some TUMMONO sequences)**  between your work and the original DSO implementation.
