### 2022-08-08 f76bb6e
adding '--net=host ' to docker run (#18)
