### 2022-08-08 bd73702
fix tum mono instructions (#19)
