### 2022-08-08 #4521 @Solonets
Hi @Mirmix, thank you for your PR!

Regarding the speed: we have the same speed with DSO on the same datasets (even though we have not yet optimized our code with SSE instructions. The test data has a bit higher image resolution, that's why it's a bit slower (DSO has the same speed on this data). You can validate that on tummono dataset, the instructions are also given in the repo.

DSO also doesn't work real time on the standard setting and uses special flags to enforce real time. Those flags are similar (but not the same) to our fast.yaml presets file
