### 2022-08-19 RoadlyInc/DSOPP#20 @jhidalgocarrio
Documentation - Energy Function, Selected Points and BA
I am missing some documentation on[ Extend DSOPP](https://github.com/RoadlyInc/DSOPP/blob/main/docs/extending_dsopp.md) about how the Energy function is implemented. I think documentation about three important points would be helpful:

* The front-end on how and where the image-to-image alignment is implemented: minimizing the photometric error from each new frame to the current (last) keyframe. Which classes are involved in the process?
* Immature points and how the implementation relates to the original DSO. Which classes are involved in this process? Where in the code, the immature points are selected to become mature points for the optimization step.
* I also miss some documentation about the Bundle Adjustment (BA). The selected keyframes to minimize the photometric error in the last N-frames. Same here, which classes are computing the BA.
