### 2022-08-19 #4521 @jhidalgocarrio
Hi @Solonets,

I see now we have live visualization of camera poses and the map. It is very helpful. Thanks.
I share the same comments with @Mirmix I don't see the GUI buttons reacting. I also get the same error when the GUI freezes.

```
$ ./build/src/application/dsopp_main --config_file_path ./test/test_data/tummono/sequence_10/standart.yaml
Framebuffer with requested attributes not available. Using available frame buffer. You may see visual artifacts
```
Another comment is speed. DSOPP performs around 10 FPS in TUMMONO, sequence 10, which is slower than in real-time.

Besides these comments, I opened an issue regarding documentation. I think this work's main advantage is demonstrating and showing to the VO community how a modern photometric image alignment is implemented. I believe this is an important aspect. The [open issue](https://github.com/RoadlyInc/DSOPP/issues/20) addresses documentation for three aspects: Energy Function, Selected Points, and BA.
