### 2022-08-25 #4521 @Solonets
Hi @Mirmix @jhidalgocarrio, thank you for your review. Right now we **do not make any performance** claims. There is a way to speed it up, and we will work in this direction later. Please make your decision based on the current version. Still, you can compile dsopp using single precision (as in dso) using the CMake flag -DUSE_FLOAT=ON. Depending on the PC it gives a performance boost from 20 to 50%.

@jhidalgocarrio I will add info about the tracker and backend into the repo.
