### 2022-08-26 377933c
Minor fixes for TUM FOV camera model
