### 2022-08-28 2a140f2
Minor fixes for TUM FOV camera model (#21)
