### 2022-08-29 05894cf
add development notes
