+++ README.md
@@ -205,7 +205,7 @@ To view it use `viewer_main` application
 
 ### [Running on own dataset](docs/running_on_own_dataset.md)
 
-### [Development notes](docs/development notes.md)
+### [Development notes](docs/development_notes.md)
 
 ### [Extending DSOPP](docs/extending_dsopp.md)
 
