### 2022-08-29 RoadlyInc/DSOPP#20 @Solonets
Hi @jhidalgocarrio. Thank you for your question. Please check this [description(https://github.com/RoadlyInc/DSOPP/tree/50054d7df1670a35dd33b95a215e89c7e316394b/docs)/development_notes.md). Does this answer your question?
