+++ .github/workflows/main.yml
@@ -1,0 +1,32 @@
+# This is a basic workflow to help you get started with Actions
+
+name: CI
+
+# Controls when the workflow will run
+on:
+  # Triggers the workflow on push or pull request events but only for the main branch
+  push:
+    branches: [ main ]
+  pull_request:
+    branches: [ main ]
+
+  # Allows you to run this workflow manually from the Actions tab
+  workflow_dispatch:
+
+jobs:
+  docker: 
+    - uses: actions/checkout@v3
+      name: Build image
+      run: docker buildx build --push \
+      --tag dsopp-latest -f $GITHUB_WORKSPACE/docker/ubuntu/Dockerfile .
+
+  download:
+    - name: Download test data
+    run: docker run -v "$GITHUB_WORKSPACE:/app" -i dsopp-latest bash -c "cd app/test; sh download_test_data.sh"
+
+  build-test:
+    runs-on: ubuntu-latest
+    name: Build & Test
+    steps:
+      - name: Build 
+        run: docker run -v "$GITHUB_WORKSPACE:/app" -i dsopp-latest bash -c "cd app;  cmake .. -DCMAKE_BUILD_TYPE=Release -DCOMPILE_HIDDEN_CODE=OFF -DLTO_OPTIMIZATION=ON -Bbuild && cmake --build build -- -j4; ctest"
