+++ .github/workflows/main.yml
@@ -17,8 +17,9 @@ jobs:
   build-gcc-test:
     runs-on: ubuntu-latest 
     steps:
-      - name: Build image
+      - name: checkout
         uses: actions/checkout@v3
+      - name: Build image
         run: docker buildx build --push --tag dsopp-latest -f $GITHUB_WORKSPACE/docker/ubuntu/Dockerfile .
       - name: Download test data
         run: docker run -v "$GITHUB_WORKSPACE:/app" -i dsopp-latest bash -c "cd app/test; sh download_test_data.sh"
