+++ test/test/energy/problems/test_affine_brightness.cpp
@@ -3,6 +3,7 @@
 #include <Eigen/Dense>
 #include <sophus/se3.hpp>
 
+#include "common/file_tools/camera_frame_times.hpp"
 #include "common/settings.hpp"
 #include "common/time/time.hpp"
 #include "energy/problems/photometric_bundle_adjustment/ceres_photometric_bundle_adjustment.hpp"
