+++ test/performance/initializer/calculate_monocular_initializer_deviations.cpp
@@ -5,6 +5,7 @@
 #include <glog/logging.h>
 #include <Eigen/Dense>
 
+#include "common/file_tools/camera_frame_times.hpp"
 #include "common/settings.hpp"
 #include "energy/camera_model/pinhole/pinhole_camera.hpp"
 #include "energy/problems/photometric_bundle_adjustment/ceres_photometric_bundle_adjustment.hpp"
