### 2022-09-03 940146f
Add exposure time to camera data frame
