+++ test/test/energy/problems/test_ceres_pose_alignment_rotated.cpp
@@ -95,8 +95,10 @@ class CeresPoseAlignmentTest : public ::testing::Test {
 
     Eigen::MatrixX<Precision> depth = readDepth(depth_img1, H, W);
 
-    auto data_frame1 = new sensors::providers::CameraDataFrame(0, std::move(img1), timeC(std::chrono::milliseconds(0)));
-    auto data_frame2 = new sensors::providers::CameraDataFrame(1, std::move(img2), timeC(std::chrono::milliseconds(1)));
+    auto data_frame1 =
+        new sensors::providers::CameraDataFrame(0, std::move(img1), 1, timeC(std::chrono::milliseconds(0)));
+    auto data_frame2 =
+        new sensors::providers::CameraDataFrame(1, std::move(img2), 1, timeC(std::chrono::milliseconds(1)));
 
     auto provider = std::make_unique<sensors::providers::MockCameraProvider>();
 
