### 2022-09-03 9098812
Add exposure time to camera features
