+++ src/tracker/depth_estimators/src/depth_estimation.cpp
@@ -38,8 +38,9 @@ void findBest(const std::vector<energy::epipolar_geometry::EpipolarLine::Epipola
               const Eigen::Vector<Precision, Pattern::kSize> &reference_pattern,
               const Eigen::Vector2<Precision> &coords, const Grid2D &target_frame, const Model &camera_model,
               const MotionProduct &t_target_reference, const sensors::calibration::CameraMask &camera_mask,
-              const Eigen::Vector2<Precision> &reference_affine_brightness,
-              const Eigen::Vector2<Precision> &target_affine_brightness, const size_t distance) {
+              const Precision reference_exposure_time, const Eigen::Vector2<Precision> &reference_affine_brightness,
+              const Precision target_exposure_time, const Eigen::Vector2<Precision> &target_affine_brightness,
+              const size_t distance) {
   const energy::reprojection::ArrayReprojector<Precision, Model, MotionProduct> reprojector(camera_model,
                                                                                             t_target_reference);
   Eigen::Matrix<Precision, 2, Pattern::kSize> reference_patch;
@@ -48,6 +49,7 @@ void findBest(const std::vector<energy::epipolar_geometry::EpipolarLine::Epipola
   Eigen::Vector<Precision, Pattern::kSize> target_pattern;
 
   const Eigen::Vector<Precision, Pattern::kSize> reference_pattern_precalc =
+      (target_exposure_time / reference_exposure_time) *
       ceres::exp(target_affine_brightness[0] - reference_affine_brightness[0]) *
       (reference_pattern - Eigen::Vector<Precision, Pattern::kSize>::Constant(reference_affine_brightness[1]));
 
@@ -79,13 +81,17 @@ class DepthEstimationProblem {
  public:
   DepthEstimationProblem(const Grid2D<C> &target_frame,
                          const Eigen::Matrix<Precision, Pattern::kSize, 1> &reference_patch,
+                         const Precision reference_exposure_time,
                          const Eigen::Vector2<Precision> &reference_affine_brightness,
+                         const Precision target_exposure_time,
                          const Eigen::Vector2<Precision> &target_affine_brightness, const Model &model,
                          Eigen::Matrix<Precision, 2, Pattern::kSize> &target_pattern,
                          const Eigen::Vector2<Precision> &tangent, const Precision sigma_huber_loss)
       : target_frame_(target_frame),
         reference_patch_(reference_patch),
+        reference_exposure_time_(reference_exposure_time),
         reference_affine_brightness_(reference_affine_brightness),
+        target_exposure_time_(target_exposure_time),
         target_affine_brightness_(target_affine_brightness),
         model_(model),
         target_pattern_(target_pattern),
@@ -93,6 +99,7 @@ class DepthEstimationProblem {
         tangent_(tangent),
         sigma_huber_loss_(sigma_huber_loss),
         reference_pattern_precalc_(
+            (target_exposure_time_ / reference_exposure_time_) *
             ceres::exp(target_affine_brightness_[0] - reference_affine_brightness_[0]) *
             (reference_patch_ - Eigen::Vector<Precision, Pattern::kSize>::Constant(reference_affine_brightness_[1]))) {}
 
@@ -156,7 +163,9 @@ class DepthEstimationProblem {
  private:
   const Grid2D<C> &target_frame_;
   const Eigen::Matrix<Precision, Pattern::kSize, 1> &reference_patch_;
+  const Precision reference_exposure_time_;
   const Eigen::Vector2<Precision> &reference_affine_brightness_;
+  const Precision target_exposure_time_;
   const Eigen::Vector2<Precision> &target_affine_brightness_;
   const Model &model_;
   Eigen::Matrix<Precision, 2, Pattern::kSize> &target_pattern_;
[…]
