### 2022-09-03 d0e2a39
Add exposure time to landmarks activator
