### 2022-09-03 b29b02d
Add exposure time to optimization
