### 2022-09-03 0eb4280
Add exposure time to keyframe strategy
