### 2022-09-04 RoadlyInc/DSOPP#20 @jhidalgocarrio
Hi @Solonets, This is great! Thanks a lot for the documentation.
Yes, the documentation is handy. I am just missing the keyframe selection strategy; which classes perform such operation?

Could you kindly add a few explanatory lines about the difference between the following classes?
* Frame
* TrackingFrame
* Keyframe
* ActiveKeyframe
