### 2022-09-04 1b5814b
Fix affine brightness test
