+++ test/test/energy/problems/test_affine_brightness.cpp
@@ -1,55 +1,52 @@
-#include <gmock/gmock.h>
 #include <gtest/gtest.h>
-#include <Eigen/Dense>
-#include <sophus/se3.hpp>
 
 #include "common/file_tools/camera_frame_times.hpp"
 #include "common/settings.hpp"
-#include "common/time/time.hpp"
-#include "energy/problems/photometric_bundle_adjustment/ceres_photometric_bundle_adjustment.hpp"
-#include "features/camera/camera_features.hpp"
-#include "features/camera/ceres_grid.hpp"
-#include "features/camera/frame_embedding_extractor.hpp"
-#include "features/camera/pattern_patch.hpp"
+#include "energy/camera_model/pinhole/pinhole_camera.hpp"
+#include "energy/motion/se3_motion.hpp"
+#include "energy/problems/photometric_bundle_adjustment/eigen_photometric_bundle_adjustment.hpp"
 #include "features/camera/pixel_data_frame.hpp"
 #include "features/camera/sobel_tracking_features_extractor.hpp"
 #include "features/camera/tracking_feature.hpp"
 #include "mock_camera_provider.hpp"
+#include "sensor/synchronized_frame.hpp"
 #include "sensors/camera/camera.hpp"
 #include "sensors/camera_calibration/fabric.hpp"
-
-#include "sensor/synchronized_frame.hpp"
-#include "sensors/camera_providers/image_folder_provider.hpp"
 #include "sensors/camera_providers/image_video_provider.hpp"
 #include "test/tools/depth_gt.hpp"
-#include "test/tools/solver_test_data.hpp"
 #include "track/active_odometry_track.hpp"
-#include "track/frames/active_keyframe.hpp"
+#include "track/active_track.hpp"
+#include "track/connections/connections_container.hpp"
+#include "track/connections/frame_connection.hpp"
 #include "track/landmarks/active_tracking_landmark.hpp"
+#include "track/landmarks/immature_tracking_landmark.hpp"
 #include "tracker/build_features.hpp"
-#include "tracker/depth_estimators/depth_estimation.hpp"
+#include "tracker/landmarks_activator/landmarks_activator.hpp"
 
-namespace dsopp {
-namespace energy {
-namespace problem {
-TEST(affine_brightness, affine_brightness) {
-  const Precision kGtA = 1;
-  const Precision kMaxAGt = 1e-8_p;
+namespace dsopp::energy::problem {
+void testAffineBrightness(bool regularize_affine_brightness, Precision exposure_time = 1) {
+  const Precision kGtA = 1.1_p;
+  const Precision kGtB = 10_p;
 
-  const Precision kGtB = 1e-4_p;
-  const Precision kMaxBGt = 1e-3_p;
+  const size_t kSensor = 0;
+  const size_t kStartFrame = 10;
+  const size_t kMaxPointsPerFrame = 800;
 
-  const size_t kFrameOffset = 1;
+  using Motion = energy::motion::SE3<Precision>;
   using Calibration = sensors::calibration::CameraCalibration;
   using Model = energy::model::PinholeCamera<Precision>;
-  static constexpr energy::model::ModelType ModelType = energy::model::ModelType::kPinholeCamera;
-  using SE3 = energy::motion::SE3<Precision>;
+  constexpr energy::model::ModelType ModelType = energy::model::ModelType::kPinholeCamera;
+
+  srand(0);
 
   auto calibration = Calibration(Eigen::Vector2<Precision>(1280, 720),
-                                 Eigen::Vector4<Precision>(448.15_p, 448.15_p, 640, 360), ModelType);
+                                 Eigen::Vector4<Precision>(448.155164329_p, 448.155164329_p, 640, 360), ModelType);
[…]
