### 2022-09-05 RoadlyInc/DSOPP#20 @Solonets
Hi @jhidalgocarrio ,

I am not particularly sure about adding those details to the documentation. You could see the class diagram and description of the classes by compiling code documentation with Doxygen. I think I just answer your question here:

Keyframe selection strategy is performed with [KeyframeStrategy](https://github.com/RoadlyInc/DSOPP/tree/main/src/tracker/keyframe_strategy/include/tracker/keyframe_strategy) interface. There are 2 implementations, you could add yours as well.

About Frames:

Frame is a base class for all frames and just carries common functionality.
ActiveKeyframe is a track's keyframes. It's position can be optimized and it carries points attached to it.
SlamIntrentalTrackingFrame is a frame that is not a keyframe. It is always attached to a keyframe and only stores its relative position to the keyframe.

Keyframe and TrackingFrame are the siblings of the 2 above classes. The only difference is that it doesn't carry additional information needed for a slam. They are used for storing and loading the track as well as manipulating it outside of the context of SLAM.
