### 2022-09-06 d8fc86a
Add affine brightness regularizers to the config file
