### 2022-09-07 afbe017
Add affine brightness regularizers to the config file (#27)
