### 2022-09-08 #4521 @hugoledoux
@Solonets I have read all the comments and re-read the updated version of the paper. I like your submission a lot, but I have a few doubts/questions:

1. **Authorship**. If I look at the commits, only 4 of the 6 authors have committed code. JOSS is primarily a software journal, and we do require authors to have contributed to the software. You can see our policy here: https://joss.readthedocs.io/en/latest/submitting.html#authorship , this means that not only commits are looked actually. Could both of you please elaborate a bit here on this issue?

2. **Contributions**: I am puzzled as to why the Python bindings that you provide are not listed as a contribution of the submission. This is an advantage over the original library. In my opinion, this ought to be emphasised more, and described in the paper.

3. **Performance**. I am a bit confused by your comments above ["There is a way to speed it up, and we will work in this direction later."](https://github.com/openjournals/joss-reviews/issues/4521#issuecomment-1227166205). Do you mean later as in the coming months or you are working on this? As @Mirmix pointed out, since you reimplement DSO and since this is a SLAM engine, I would expect some form of performances. It does not need to be faster, but being slower would be pretty bad, wouldn't it? You claim that it's as fast, but I think this should be reported and benchmarked in the readme/paper.
