### 2022-09-09 #4521 @Solonets
Hi @hugoledoux. Thank you for your feedback. I will answer your questions one by one.

1. We developed the code before in the private repository.  I later removed some modules and submitted the code as one commit to the new repository.  All the people in the title except for one (CEO) heavily contributed to the code. After reading the authorship policy I agree we have to remove him from the authors.

2. There are no really extensive bindings in the code. There is a python library to read the resulting pose graph and some rare bindings to some functionality, but that's it.

3. The dso code is extremely optimized with SSE instructions.  It's a lot of work to be done to speed up each module with them. I am currently looking for students at TUM to take over this work. Still, on my computer with similar settings, the speed is very close to the original dso.  ![fps](https://user-images.githubusercontent.com/3417600/189332842-187569c6-5126-49b1-ada0-7b4503f6819a.png)
