### 2022-09-09 dc4b900
Export points in same coordinate system
