### 2022-09-17 RoadlyInc/DSOPP#20 @jhidalgocarrio
As you prefer.
I still think a class diagram is a description of all the classes. While this [doc](https://github.com/RoadlyInc/DSOPP/blob/50054d7df1670a35dd33b95a215e89c7e316394b/docs/development_notes.md) is only about the key parts of DSO in order to quickly grasp the core of DSO in your implementation.

Just adding the text from your reply would help:

```
Frame is a base class for all frames and just carries common functionality.
ActiveKeyframe is a track's keyframes. Its position can be optimized, and it carries points attached to it.
SlamIntrentalTrackingFrame is a frame that is not a keyframe. It is always attached to a keyframe and only stores its relative position to the keyframe.

Keyframe and TrackingFrame are the siblings of the 2 above classes. The only difference is that it doesn't carry additional information needed for a slam. They are used for storing and loading the track as well as manipulating it outside of the context of SLAM.
```
