### 2022-12-19 c715b9d
Add PixelMap<C> object pool
