+++ pydsopp/track/track.py
@@ -30,6 +30,7 @@ class Track:
     """
     Class for exctract data from track storage
     """
+
     def __init__(self, track_storage):
         """
         Arguments:
