+++ pydsopp/track/export/colmap_track_export.py
@@ -12,6 +12,7 @@ class Point3D:
     """
     Internal class for building colmap data, representation of 3D point with index and 2D projection indices
     """
+
     @dataclass
     class Projection:
         frame_index: int
@@ -31,6 +32,7 @@ class Frame:
     """
     Internal class for building colmap data
     """
+
     @dataclass
     class Point2D:
         pixel: np.ndarray
