+++ src/features/src/pixel_map.cpp
@@ -1,14 +1,107 @@
 
 #include "features/camera/pixel_map.hpp"
 
+#include <mutex>
+#include <stack>
+
 #include <glog/logging.h>
 
 namespace dsopp {
 namespace features {
 
+/**
+ * object pool that caches the underlying data storage for computed `PixelMap`s the pool caches storage containers
+ * based on their size to avoid performing resizes on them
+ */
+template <int C>
+class PixelInfoPool {
+  /** type of the container used to store PixelInfo data */
+  using PixelInfoContainer = std::vector<PixelInfo<C>, typename PixelInfo<C>::PixelInfoAllocator>;
+
+ private:
+  /** custom deleter for shared pointers, returns the data storage container back into its original pool */
+  struct ExternalDeleter {
+    /**
+     * creates a custom deleter object
+     * @param pool reference to the pool that the deleted container should be placed into
+     */
+    explicit ExternalDeleter(std::weak_ptr<PixelInfoPool *> pool) : pool_(pool) {}
+
+    /**
+     * called after the last reference to a shared_ptr dies
+     * @param ptr pointer to the container instance that the shared_ptr managed
+     */
+    void operator()(PixelInfoContainer *ptr) {
+      if (auto pool_ptr = pool_.lock()) {
+        try {
+          (*pool_ptr.get())->add(std::unique_ptr<PixelInfoContainer>{ptr}, ptr->capacity());
+          return;
+        } catch (...) {
+        }
+      }
+      std::default_delete<PixelInfoContainer>{}(ptr);
+    }
+
+   private:
+    /** reference to the pool that created this deleter */
+    std::weak_ptr<PixelInfoPool *> pool_;
+  };
+
+ public:
+  PixelInfoPool() : this_ptr_(new PixelInfoPool *(this)) {}
+  virtual ~PixelInfoPool() {}
+
+  /**
+   * add the given storage container into the pool with a given size
+   * @param t container to be placed into the pool
+   * @param container_size number of allocated elements in the given container
+   */
+  void add(std::unique_ptr<PixelInfoContainer> t, const size_t container_size) {
+    std::lock_guard<std::mutex> lock(mutex_);
+    pools_[container_size].push(std::move(t));
+  }
+
+  /**
+   * get a storage container of a desired size from the pool
+   * @param desired_size number of elements the container should have allocated
+   * @return a shared pointer to a container of requested size
+   */
+  std::shared_ptr<PixelInfoContainer> acquire(const size_t desired_size) {
+    std::lock_guard<std::mutex> lock(mutex_);
+    std::stack<std::unique_ptr<PixelInfoContainer>> &pool = pools_[desired_size];
+    if (pool.empty()) {
+      // if no unused containers are available, create and store a new one
+      std::unique_ptr<PixelInfoContainer> t(new PixelInfoContainer());
+      t->resize(desired_size);
+      pool.push(std::move(t));
+    }
+
+    // create a shared_ptr with a custom deleter, that will automatically return the container back to this pool
+    std::shared_ptr<PixelInfoContainer> tmp(pool.top().release(),
+                                            ExternalDeleter{std::weak_ptr<PixelInfoPool *>{this_ptr_}});
+    pool.pop();
+    return tmp;
+  }
+
+ private:
+  /** mutex to share the pool between threads */
+  std::mutex mutex_;
+  /** pointer to this pool, used to create shared_ptr deleters */
+  std::shared_ptr<PixelInfoPool *> this_ptr_;
+  /** this map stores multiple object pools according to the capacity of the containers for each of them */
[…]
