### 2022-12-20 9f9f685
Accelerate pixel map
