+++ src/features/src/pixel_map.cpp
@@ -1,5 +1,6 @@
 
 #include "features/camera/pixel_map.hpp"
+#include "features/camera/calculate_pixelinfo.hpp"
 
 #include <mutex>
 #include <stack>
@@ -93,7 +94,8 @@ class PixelInfoPool {
 };
 
 template <int C>
-PixelMap<C>::PixelMap(std::vector<Precision> &&data, long width, long height) : plain_data_(std::move(data)) {
+PixelMap<C>::PixelMap(std::vector<Precision, PrecisionAllocator> &&data, long width, long height)
+    : plain_data_(std::move(data)) {
   CHECK_EQ(plain_data_.size(), width * height * C) << "Incorrect size of vector of data";
 
   // note: initialization of a static local variable is supposed to be thread-safe according to
@@ -103,39 +105,8 @@ PixelMap<C>::PixelMap(std::vector<Precision> &&data, long width, long height) : 
   pixelinfo_data_ = pool_.acquire(static_cast<size_t>(height * width));
   new (&map_) PixelInfoStorage(pixelinfo_data_->data(), height, width);
 
-  using Channel = Eigen::Map<Eigen::Matrix<Precision, -1, -1, Eigen::RowMajor>>;
-
-  map_.resize(height, width);
-
-  for (int i = 0; i < C; ++i) {
-    Channel channel(&plain_data_[static_cast<size_t>(i * width * height)], height, width);
-
-    for (long y = 0; y < height; ++y)
-      for (long x = 0; x < width; ++x) {
-        Precision c = channel(y, x);
-        map_(y, x).data_(i, 0) = c;
-
-        Precision dx;
-        if (x == 0) {
-          dx = channel(y, x + 1) - channel(y, x);
-        } else if (x == width - 1) {
-          dx = channel(y, x) - channel(y, x - 1);
-        } else {
-          dx = 0.5_p * (channel(y, x + 1) - channel(y, x - 1));
-        }
-        map_(y, x).data_(i, 1) = dx;
-
-        Precision dy;
-        if (y == 0) {
-          dy = channel(y + 1, x) - channel(y, x);
-        } else if (y == height - 1) {
-          dy = channel(y, x) - channel(y - 1, x);
-        } else {
-          dy = 0.5_p * (channel(y + 1, x) - channel(y - 1, x));
-        }
-        map_(y, x).data_(i, 2) = dy;
-      }
-  }
+  calculate_pixelinfo<C>(plain_data_.data(), static_cast<Precision *>(static_cast<void *>(map_.data())),
+                         static_cast<int>(width), static_cast<int>(height));
 }
 
 template <int C>
@@ -173,7 +144,7 @@ size_t PixelMap<C>::size() const {
 }
 
 template <int C>
-const std::vector<Precision> &PixelMap<C>::data() const {
+const std::vector<Precision, PrecisionAllocator> &PixelMap<C>::data() const {
   return plain_data_;
 }
 template <int C>
