+++ src/features/include/features/camera/pixel_map.hpp
@@ -152,7 +152,7 @@ class PixelMap {
    * @param width width of the map
    * @param height height of the map
    */
-  PixelMap(std::vector<Precision> &&data, long width, long height);
+  PixelMap(std::vector<Precision, PrecisionAllocator> &&data, long width, long height);
   /** default constructor */
   PixelMap();
   /** default move constructor */
@@ -190,7 +190,7 @@ class PixelMap {
    *
    * @return full data from the map
    * */
-  const std::vector<Precision> &data() const;
+  const std::vector<Precision, PrecisionAllocator> &data() const;
 
   /**
    * method to get PixelInfo for individual pixel
@@ -322,7 +322,7 @@ class PixelMap {
 
  private:
   /** Container with all image pixels [channel][height][width]  */
-  std::vector<Precision> plain_data_;
+  std::vector<Precision, PrecisionAllocator> plain_data_;
 
   /** Container with channel info for all pixels [height][width][PixelInfo<C>]  */
   std::shared_ptr<std::vector<PixelInfo<C>, typename PixelInfo<C>::PixelInfoAllocator>> pixelinfo_data_;
