### 2022-12-21 a4af2aa
Accelerate pixel map (#29)
